agents/orchestration_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
aiohttp==3.9.1
httpx==0.25.2
pyyaml==6.0.1
python-multipart==0.0.6
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==0.42b0
pika==1.3.2
redis==5.0.1
tenacity==8.2.3
python-json-logger==2.0.7